backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
nbformat==5.9.2
nbconvert==7.11.0
mistune<3.0  # Fix for nbconvert compatibility
google-generativeai==0.3.1
openai==1.3.7
python-dotenv==1.0.0
